Extract an exception handler and move errorResponse generation there.

rh-pre-commit.version: 2.0.1
rh-pre-commit.check-secrets: ENABLED
